abstraction over simulator cyclers
