+++ circuit_defs/reduced_risc_v_machine/verifier/src/tests.rs
@@ -259,11 +259,11 @@ fn test_verifier_in_simulator() {
 
         let inner = VecDeque::<u32>::from(oracle_data);
         let oracle = VectorBasedNonDeterminismSource(inner, QuasiUARTSourceState::Ready);
-        let (_, output) = run_simple_with_entry_point_and_non_determimism_source_for_config::<
+        let output = run_simple_with_entry_point_and_non_determimism_source_for_config::<
             _,
             IWithoutByteAccessIsaConfigWithDelegation,
             // IMIsaConfigWithAllDelegations,
         >(config, oracle);
-        dbg!(output);
+        dbg!(output.state);
     }
 }
