+++ prover/src/tracers/main_cycle_optimized.rs
@@ -426,7 +426,7 @@ impl<
                     .per_cycle_data
                     .last_mut()
                     .unwrap_unchecked()
-                    .pc = current_state.pc;
+                    .pc = current_state.observable.pc;
             }
         }
     }
