+++ risc_v_simulator/src/delegations/blake2_round_function_with_compression_mode.rs
@@ -32,6 +32,8 @@ pub fn blake2_round_function_with_extended_control<
 ) {
     assert_eq!(rs1_value, 0, "aligned memory access is unused");
 
+    let state = state.observable;
+
     // read registers first
     let x10 = state.registers[10];
     let x11 = state.registers[11];
