+++ risc_v_simulator/src/tests/mod.rs
@@ -25,8 +25,9 @@ fn test_reg_reg_op(op_name: &str, expected: u32, op1: u32, op2: u32) {
     {
         // old simulator
         let mut state = RiscV32State::<IMStandardIsaConfig>::initial(INITIAL_PC);
-        state.registers[1] = op1;
-        state.registers[2] = op2;
+        
+        state.observable.registers[1] = op1;
+        state.observable.registers[2] = op2;
         let instr = format!("{} x3, x1, x2", op_name);
         let mut empty_hash: HashMap<String, u32> = HashMap::new();
         let encoding = lib_rv32_asm::assemble_ir(&instr, &mut empty_hash, INITIAL_PC)
@@ -42,7 +43,7 @@ fn test_reg_reg_op(op_name: &str, expected: u32, op1: u32, op2: u32) {
         }
         let mut mmu = NoMMU::default();
         state.cycle(&mut memory, &mut (), &mut mmu, &mut ZeroedSource);
-        assert!(state.registers[3] == expected, "Unexpected output: expected 0x{:08x} for operation `{}` 0x{:08x}, 0x{:08x}, obtained 0x{:08x}", expected, op_name, op1, op2, state.registers[3]);
+        assert!(state.observable.registers[3] == expected, "Unexpected output: expected 0x{:08x} for operation `{}` 0x{:08x}, 0x{:08x}, obtained 0x{:08x}", expected, op_name, op1, op2, state.observable.registers[3]);
     }
     {
         // new simulator
@@ -70,7 +71,7 @@ fn test_reg_reg_op(op_name: &str, expected: u32, op1: u32, op2: u32) {
 fn test_reg_imm_op(op_name: &str, expected: u32, op1: u32, imm: u16) {
     {
         let mut state = RiscV32State::<IMStandardIsaConfig>::initial(INITIAL_PC);
-        state.registers[1] = op1;
+        state.observable.registers[1] = op1;
         let instr = format!("{} x3, x1, 0x{:x}", op_name, imm);
         let mut empty_hash: HashMap<String, u32> = HashMap::new();
         let encoding = lib_rv32_asm::assemble_ir(&instr, &mut empty_hash, INITIAL_PC)
@@ -86,7 +87,7 @@ fn test_reg_imm_op(op_name: &str, expected: u32, op1: u32, imm: u16) {
         }
         let mut mmu = NoMMU::default();
         state.cycle(&mut memory, &mut (), &mut mmu, &mut ZeroedSource);
-        assert!(state.registers[3] == expected, "Unexpected output: expected 0x{:08x} for operation `{}` 0x{:08x}, 0x{:04x}, obtained 0x{:08x}", expected, op_name, op1, imm, state.registers[3]);
+        assert!(state.observable.registers[3] == expected, "Unexpected output: expected 0x{:08x} for operation `{}` 0x{:08x}, 0x{:04x}, obtained 0x{:08x}", expected, op_name, op1, imm, state.observable.registers[3]);
     }
     {
         let mut state = RiscV32StateForUnrolledProver::<IMStandardIsaConfig>::initial(INITIAL_PC);
@@ -112,8 +113,8 @@ fn test_reg_imm_op(op_name: &str, expected: u32, op1: u32, imm: u16) {
 fn test_branch_op<const TAKEN: bool>(op_name: &str, op1: u32, op2: u32) {
     {
         let mut state = RiscV32State::<IMStandardIsaConfig>::initial(INITIAL_PC);
-        state.registers[1] = op1;
-        state.registers[2] = op2;
+        state.observable.registers[1] = op1;
+        state.observable.registers[2] = op2;
         let instr = format!("{} x1, x2, 0x08", op_name);
         let mut empty_hash: HashMap<String, u32> = HashMap::new();
         let encoding = lib_rv32_asm::assemble_ir(&instr, &mut empty_hash, INITIAL_PC)
@@ -129,17 +130,17 @@ fn test_branch_op<const TAKEN: bool>(op_name: &str, op1: u32, op2: u32) {
         }
         let mut mmu = NoMMU::default();
         state.cycle(&mut memory, &mut (), &mut mmu, &mut ZeroedSource);
-        assert!(state.pc == 4 || state.pc == 8);
+        assert!(state.observable.pc == 4 || state.observable.pc == 8);
         if TAKEN {
             assert!(
-                state.pc == 8,
+                state.observable.pc == 8,
                 "Unexpected branching: expected to take a branch for operation `{}` 0x{:08x}, 0x{:08x}",
                 op_name,
                 op1,
                 op2
             );
         } else {
-            assert!(state.pc == 4, "Unexpected branching: expected to NOT take a branch for operation `{}` 0x{:08x}, 0x{:08x}", op_name, op1, op2);
+            assert!(state.observable.pc == 4, "Unexpected branching: expected to NOT take a branch for operation `{}` 0x{:08x}, 0x{:08x}", op_name, op1, op2);
         }
     }
     {
